Tests/LibWeb/Text/input/async-scrolling/main-thread-scroll-rebase.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style>
    body {
        margin: 0;
    }

    #scroller {
        width: 100px;
        height: 100px;
        overflow: scroll;
    }

    #spacer {
        height: 600px;
    }
</style>
<div id="scroller"><div id="spacer"></div></div>
<script>
    promiseTest(async () => {
        await animationFrame();
        await animationFrame();

        const wheelPromise = internals.wheel(50, 50, 0, 100);
        scroller.scrollTop = 200;
        println(`script scrollTop: ${scroller.scrollTop}`);
        scroller.appendChild(document.createElement("div"));

        await wheelPromise;

        println(`final scrollTop: ${scroller.scrollTop}`);
    });
</script>
