Tests/LibWeb/Text/expected/async-scrolling/main-thread-scroll-rebase.txt

script scrollTop: 200
final scrollTop: 300
